New build script.

+++ camera.cpp
@@ -16,12 +16,12 @@ void recalc_view_matrix(Camera* c)
 void init(Camera* c)
 {
     memset(c, 0, sizeof(Camera));
-    float near_plane = 0.01f;
-    float far_plane = 1000.0f;
-    float fov = 90.0f;
-    float aspect = 1.0f;
-    float y_scale = 1.0f / tan((3.14f / 180.0f) * fov / 2);
-    float x_scale = y_scale / aspect;
+    f32 near_plane = 0.01f;
+    f32 far_plane = 1000.0f;
+    f32 fov = 90.0f;
+    f32 aspect = 1.0f;
+    f32 y_scale = 1.0f / tanf((3.14f / 180.0f) * fov / 2);
+    f32 x_scale = y_scale / aspect;
     c->projection_matrix = {
         x_scale, 0, 0, 0,
         0, y_scale, 0, 0,
